+++ processor/src/test/fixtures/expected/com/example/arez/memoize/React4j_BasicMemoizeComponent.java
@@ -11,7 +11,6 @@ import arez.annotations.InjectMode;
 import arez.annotations.Observe;
 import arez.annotations.ObserverRef;
 import arez.annotations.Priority;
-import arez.annotations.PriorityOverride;
 import javax.annotation.Generated;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
@@ -34,6 +33,7 @@ import react4j.internal.arez.SchedulerUtil;
 @ArezComponent(
     name = "BasicMemoizeComponent",
     disposeNotifier = Feature.DISABLE,
+    defaultPriority = Priority.LOWEST,
     inject = InjectMode.NONE
 )
 @Generated("react4j.processor.ReactProcessor")
@@ -113,11 +113,6 @@ abstract class React4j_BasicMemoizeComponent extends BasicMemoizeComponent {
     if ( React.shouldStoreDebugDataAsState() && Arez.areSpiesEnabled() ) {
       IntrospectUtil.collectDependencyDebugData( getRenderObserver(), data );
     }
-  }
-
-  @PriorityOverride
-  final int iconPriority() {
-    return Observer.Flags.PRIORITY_LOWEST;
   }
 
   static final class Factory {