+++ processor/src/test/fixtures/expected/com/example/arez/memoize/React4j_PropAndMemoizeComponent.java
@@ -15,7 +15,6 @@ import arez.annotations.ObservableValueRef;
 import arez.annotations.Observe;
 import arez.annotations.ObserverRef;
 import arez.annotations.Priority;
-import arez.annotations.PriorityOverride;
 import javax.annotation.Generated;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
@@ -38,6 +37,7 @@ import react4j.internal.arez.SchedulerUtil;
 @ArezComponent(
     name = "PropAndMemoizeComponent",
     disposeNotifier = Feature.DISABLE,
+    defaultPriority = Priority.LOWEST,
     inject = InjectMode.NONE
 )
 @Generated("react4j.processor.ReactProcessor")
@@ -63,7 +63,7 @@ abstract class React4j_PropAndMemoizeComponent extends PropAndMemoizeComponent {
       expectSetter = false,
       readOutsideTransaction = true
   )
-  protected String getValue() {
+  String getValue() {
     if ( React.shouldCheckInvariants() ) {
       return null != props().getAsAny( Props.value ) ? props().getAsAny( Props.value ).asString() : null;
     } else {
@@ -143,11 +143,6 @@ abstract class React4j_PropAndMemoizeComponent extends PropAndMemoizeComponent {
     if ( React.shouldStoreDebugDataAsState() && Arez.areSpiesEnabled() ) {
       IntrospectUtil.collectDependencyDebugData( getRenderObserver(), data );
     }
-  }
-
-  @PriorityOverride
-  final int activePriority() {
-    return Observer.Flags.PRIORITY_LOWEST;
   }
 
   static final class Factory {