+++ processor/src/test/fixtures/expected/com/example/prop/React4j_ObservableProp.java
@@ -62,7 +62,7 @@ abstract class React4j_ObservableProp extends ObservableProp {
       expectSetter = false,
       readOutsideTransaction = true
   )
-  protected Object getValue() {
+  Object getValue() {
     if ( React.shouldCheckInvariants() ) {
       return null != props().getAsAny( Props.value ) ? props().getAsAny( Props.value ).cast() : null;
     } else {