+++ processor/src/test/fixtures/expected/com/example/prop/React4j_ObservableViaObservedProp.java
@@ -47,7 +47,7 @@ abstract class React4j_ObservableViaObservedProp extends ObservableViaObservedPr
       expectSetter = false,
       readOutsideTransaction = true
   )
-  protected Object getValue() {
+  Object getValue() {
     if ( React.shouldCheckInvariants() ) {
       return null != props().getAsAny( Props.value ) ? props().getAsAny( Props.value ).cast() : null;
     } else {