+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Unreleased
 
 * Upgrade the `org.realityforge.arez` artifacts to version `0.155`.
+* If a React4j component has a `@Memoize` or `@Observe` annotated method other than render then specify the `defaultPriority` parameter to `Priority.LOWEST` on the `@ArezComponent` annotation in the enhanced class. The replaces the usage of the `@PriorityOveride` annotation from Arez that has now been eliminated. This change also means that the default priority for `@Observe` annotated methods is also `LOWEST`. 
 
 ### [v0.141](https://github.com/react4j/react4j/tree/v0.141) (2019-11-12) · [Full Changelog](https://github.com/react4j/react4j/compare/v0.140...v0.141)